+++ src/HdbConfigurationManager.cpp
@@ -309,14 +309,14 @@ void HdbConfigurationManager::init_device()
 	
 	//	Initialize device
 	*attr_SetCodePushedEvent_read = false;
-	*attr_SetAttributeName_read = CORBA::string_dup("");
-	*attr_SetArchiver_read = CORBA::string_dup("");
+	*attr_SetAttributeName_read = Tango::string_dup("");
+	*attr_SetArchiver_read = Tango::string_dup("");
 	*attr_SetPeriodEvent_read = -1;
 	*attr_SetRelativeEvent_read = -1;
 	*attr_SetAbsoluteEvent_read = -1;
 	*attr_SetPollingPeriod_read = -1;
 	*attr_SetTTL_read = 0;
-	*attr_SetStrategy_read = CORBA::string_dup("");
+	*attr_SetStrategy_read = Tango::string_dup("");
 	archiver_t tmp;
 	for(auto it = archiverList.begin(); it!= archiverList.end(); it++)
 	{
@@ -858,7 +858,8 @@ void HdbConfigurationManager::write_SetAttributeName(Tango::WAttribute &attr)
 					(const char *)__func__);
 	}
 	delete dp;
-	*attr_SetAttributeName_read = CORBA::string_dup(signame.c_str());
+	Tango::string_free(*attr_SetAttributeName_read);
+	*attr_SetAttributeName_read = Tango::string_dup(signame.c_str());
 	/*----- PROTECTED REGION END -----*/	//	HdbConfigurationManager::write_SetAttributeName
 }
 //--------------------------------------------------------
@@ -1127,7 +1128,8 @@ void HdbConfigurationManager::write_SetArchiver(Tango::WAttribute &attr)
 	/*----- PROTECTED REGION ID(HdbConfigurationManager::write_SetArchiver) ENABLED START -----*/
 	string signame(w_val);
 	fix_tango_host(signame);
-	*attr_SetArchiver_read = CORBA::string_dup(signame.c_str());
+	Tango::string_free(*attr_SetArchiver_read);
+	*attr_SetArchiver_read = Tango::string_dup(signame.c_str());
 	
 	/*----- PROTECTED REGION END -----*/	//	HdbConfigurationManager::write_SetArchiver
 }
@@ -1583,7 +1585,8 @@ void HdbConfigurationManager::write_SetStrategy(Tango::WAttribute &attr)
 	Tango::DevString	w_val;
 	attr.get_write_value(w_val);
 	/*----- PROTECTED REGION ID(HdbConfigurationManager::write_SetStrategy) ENABLED START -----*/
-	*attr_SetStrategy_read = CORBA::string_dup(w_val);
+	Tango::string_free(*attr_SetStrategy_read);
+	*attr_SetStrategy_read = Tango::string_dup(w_val);
 	/*----- PROTECTED REGION END -----*/	//	HdbConfigurationManager::write_SetStrategy
 }
 //--------------------------------------------------------
@@ -1952,12 +1955,12 @@ void HdbConfigurationManager::attribute_add()
 	//------2: set attribute with event parameters-------------------------
 
 	vector<string> attrnames;
+	Tango::AttributeInfoListEx *attr_info_list_ex = nullptr;
 	try
 	{
 		attrnames.push_back(attname);
 
 		//read actual attribute config
-		Tango::AttributeInfoListEx *attr_info_list_ex;
 		DEBUG_STREAM << "HdbConfigurationManager::AttributeAdd()  - before read attribute config attr="<<attname;
 		attr_info_list_ex = dp->get_attribute_config_ex(attrnames);
 		DEBUG_STREAM << "HdbConfigurationManager::AttributeAdd()  - read attribute config size=" << attr_info_list_ex->size() << endl;
@@ -1992,8 +1995,7 @@ void HdbConfigurationManager::attribute_add()
 		}
 		if(changed)
 			dp->set_attribute_config(*attr_info_list_ex);
-
-
+		
 		DEBUG_STREAM<<"after  set_attribute_config";
 		//write polling only if changed
 		if(*attr_SetPollingPeriod_read != 0 && original_SetPollingPeriod != *attr_SetPollingPeriod_read)
@@ -2012,6 +2014,7 @@ void HdbConfigurationManager::attribute_add()
 	{
 	}
 	delete dp;
+	delete attr_info_list_ex;
 	//------3: Configure DB------------------------------------------------
 	/*
 	try
@@ -2410,13 +2413,13 @@ void HdbConfigurationManager::attribute_assign(const Tango::DevVarStringArray *a
 		else
 			add_argin.length(2);
 
-		add_argin[0] = CORBA::string_dup(signame.c_str());
-		add_argin[1] = CORBA::string_dup(strategy.c_str());
+		add_argin[0] = Tango::string_dup(signame.c_str());
+		add_argin[1] = Tango::string_dup(strategy.c_str());
         
 		if (argin->length() == 4)
 		{
 			string ttl((*argin)[3]);
-			add_argin[2] = CORBA::string_dup(ttl.c_str());
+			add_argin[2] = Tango::string_dup(ttl.c_str());
 		}
 
 		Din << add_argin;
@@ -2486,7 +2489,7 @@ Tango::DevString HdbConfigurationManager::attribute_status(Tango::DevString argi
 				Dout = itmap->second.dp->command_inout("AttributeStatus",Din);
 				Dout >> status;
 				status += string("\nArchiver: ") + itmap->first;
-				argout = CORBA::string_dup(status.c_str());
+				argout = Tango::string_dup(status.c_str());
 				found = true;
 				break;
 			}
@@ -2536,7 +2539,7 @@ Tango::DevString HdbConfigurationManager::attribute_get_archiver(Tango::DevStrin
 					(const char*)__func__, Tango::ERR);
 	}*/
 	
-	argout = CORBA::string_dup(archiver.c_str());
+	argout = Tango::string_dup(archiver.c_str());
 	/*----- PROTECTED REGION END -----*/	//	HdbConfigurationManager::attribute_get_archiver
 	return argout;
 }
@@ -2612,7 +2615,7 @@ Tango::DevVarStringArray *HdbConfigurationManager::attribute_search(Tango::DevSt
 	for (vector<string>::iterator it= attribute_search_list_str.begin(); it != attribute_search_list_str.end(); it++)
 	{
 		DEBUG_STREAM << "HdbConfigurationManager::AttributeSearch()  - result list -> copying=" << *it << endl;
-		(*argout)[i] = CORBA::string_dup(it->c_str());
+		(*argout)[i] = Tango::string_dup(it->c_str());
 		i++;
 	}
 	
@@ -2800,8 +2803,8 @@ void HdbConfigurationManager::set_attribute_strategy(const Tango::DevVarStringAr
 		Tango::DevVarStringArray add_argin;
 		Tango::DeviceData Din;
 		add_argin.length(2);
-		add_argin[0] = CORBA::string_dup(signame.c_str());
-		add_argin[1] = CORBA::string_dup(v_contexts.c_str());
+		add_argin[0] = Tango::string_dup(signame.c_str());
+		add_argin[1] = Tango::string_dup(v_contexts.c_str());
 		Din << add_argin;
 		itmap->second.dp->command_inout("SetAttributeStrategy",Din);
 	}
@@ -2856,11 +2859,11 @@ Tango::DevString HdbConfigurationManager::get_attribute_strategy(Tango::DevStrin
 		Din << signame;
 		Dout = itmap->second.dp->command_inout("GetAttributeStrategy",Din);
 		Dout >> strategy;
-		argout = CORBA::string_dup(strategy.c_str());
+		argout = Tango::string_dup(strategy.c_str());
 	}
 	else
 	{
-		argout = CORBA::string_dup("");
+		argout = Tango::string_dup("");
 	}
 
 	
@@ -2912,8 +2915,8 @@ void HdbConfigurationManager::set_attribute_ttl(const Tango::DevVarStringArray *
 		Tango::DevVarStringArray add_argin;
 		Tango::DeviceData Din;
 		add_argin.length(2);
-		add_argin[0] = CORBA::string_dup(signame.c_str());
-		add_argin[1] = CORBA::string_dup(s_ttl.c_str());
+		add_argin[0] = Tango::string_dup(signame.c_str());
+		add_argin[1] = Tango::string_dup(s_ttl.c_str());
 		Din << add_argin;
 		itmap->second.dp->command_inout("SetAttributeTTL",Din);
 	}
@@ -3070,20 +3073,29 @@ void HdbConfigurationManager::add_domain(string &str)
 		hints.ai_family = AF_UNSPEC; /*either IPV4 or IPV6*/
 		hints.ai_socktype = SOCK_STREAM;
 		hints.ai_flags = AI_CANONNAME;
-		struct addrinfo *result, *rp;
+		struct addrinfo *result;
 		int ret = getaddrinfo(th.c_str(), NULL, &hints, &result);
 		if (ret != 0)
 		{
 			cout << __func__<< ": getaddrinfo error='" << gai_strerror(ret)<<"' while looking for " << th<<endl;
 			return;
 		}
-
-		for (rp = result; rp != NULL; rp = rp->ai_next)
+		if(result == NULL)
 		{
-			with_domain = string(rp->ai_canonname) + str.substr(end2);
-			//cout << __func__ <<": found domain -> " << with_domain<<endl;
-			domain_map.insert(make_pair(th, with_domain));
+			cout << __func__ << ": getaddrinfo did not return domain information for " 
+			     << th << " (result == NULL)" << endl;
+			return;
 		}
+		if (result->ai_canonname == NULL)
+		{
+			cout << __func__ << ": getaddrinfo did not return domain information for " 
+			     << th << " (result->ai_canonname == NULL)" << endl;
+			freeaddrinfo(result);
+			return;
+		}
+		with_domain = string(result->ai_canonname) + str.substr(end2);
+		//cout << __func__ <<": found domain -> " << with_domain<<endl;
+		domain_map.insert(make_pair(th, with_domain));
 		freeaddrinfo(result); // all done with this structure
 		str = with_domain;
 		return;
@@ -3152,7 +3164,7 @@ void HdbConfigurationManager::add_domain(string &str)
 			hints.ai_family = AF_UNSPEC; /*either IPV4 or IPV6*/
 			hints.ai_socktype = SOCK_STREAM;
 			hints.ai_flags = AI_CANONNAME;
-			struct addrinfo *result, *rp;
+			struct addrinfo *result;
 			int ret = getaddrinfo(th.c_str(), NULL, &hints, &result);
 			if (ret != 0)
 			{
@@ -3162,12 +3174,21 @@ void HdbConfigurationManager::add_domain(string &str)
 					strresult += ",";
 				continue;
 			}
-
-			for (rp = result; rp != NULL; rp = rp->ai_next)
+			if(result == NULL)
 			{
-				with_domain = string(rp->ai_canonname) + it->substr(end2);
-				domain_map.insert(make_pair(th, string(rp->ai_canonname)));
+				cout << __func__ << ": getaddrinfo did not return the canonical name for " 
+				     << th << " (result == NULL)" << endl;
+				return;
 			}
+			if (result->ai_canonname == NULL)
+			{
+				cout << __func__ << ": getaddrinfo did not return the canonical name for " 
+				     << th << " (result->ai_canonname == NULL)" << endl;
+				freeaddrinfo(result);
+				return;
+			}
+			with_domain = string(result->ai_canonname) + it->substr(end2);
+			domain_map.insert(make_pair(th, string(rp->ai_canonname)));
 			freeaddrinfo(result); // all done with this structure
 			strresult += with_domain;
 			if(it != facilities.end()-1)